+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Release history
 
-* **0.02** (2014/07/29) Integrated BLAS and LAPACK. 
+* **0.03** (2014/08/12) Integrated [LAPACKE](http://www.netlib.org/lapack/lapacke.html) because the existing CLAPACK wrapper did not support row-major order.
 
-	To do: incorporate unit tests based on examples from [NAG](http://www.nag.com/lapack-ex/lapack-ex.html).
+* **0.02** (2014/07/29) Integrated BLAS and LAPACK. 
 
 * **0.01** (2014/07/21) Initial release.